Move FrontendApplicationContribution into own file
