+++ packages/core/src/browser/dialogs.ts
@@ -18,7 +18,7 @@ import { injectable, inject } from 'inversify';
 import { Disposable, MaybePromise, CancellationTokenSource, nls } from '../common';
 import { Key } from './keyboard/keys';
 import { Widget, BaseWidget, Message, addKeyListener, codiconArray } from './widgets';
-import { FrontendApplicationContribution } from './frontend-application';
+import { FrontendApplicationContribution } from './frontend-application-contribution';
 
 @injectable()
 export class DialogProps {
