+++ packages/core/src/browser/quick-input/quick-input-frontend-contribution.ts
@@ -16,7 +16,7 @@
 
 import { injectable, inject, named } from 'inversify';
 import { ContributionProvider } from '../../common';
-import { FrontendApplicationContribution } from '../frontend-application';
+import { FrontendApplicationContribution } from '../frontend-application-contribution';
 import { QuickAccessContribution } from './quick-access';
 
 @injectable()
