+++ packages/core/src/browser/window/default-window-service.spec.ts
@@ -17,7 +17,7 @@
 import { Container } from 'inversify';
 import { ContributionProvider } from '../../common';
 import { CorePreferences } from '../core-preferences';
-import { FrontendApplicationContribution } from '../frontend-application';
+import { FrontendApplicationContribution } from '../frontend-application-contribution';
 import { DefaultWindowService } from './default-window-service';
 import assert = require('assert');
